Feat: update claim design page - create 3 seperate page componenent
